Add files via upload
In this project, I have created two different files  `movies` & `ratings`, and populate the CSV data into them. (file is in config folder). And then crated a Get route which returns as JSON, the top 10 movies with the longest runTime. And  then I created a Post route which takes JSON as input for new movie and saves it into the database.